alignnG

+++ matdeeplearn/models/alignn_graphite.py
@@ -25,6 +25,7 @@ class ALIGNN_GRAPHITE(BaseModel):
         self.num_interactions = alignn_layers
         self.cutoff = max_edge_distance
 
+        #self.embed_atm = Embedding(data.num_features, hidden_features)
         self.embed_atm = Linear(data.num_features, hidden_features)
         self.embed_bnd = partial(bessel, start=0, end=max_edge_distance, num_basis=hidden_features)
 
@@ -64,7 +65,7 @@ class ALIGNN_GRAPHITE(BaseModel):
         edge_index_A = data.edge_index_lg
         print(data.x.size())
         print(data.x)
-        h_atm = self.embed_atm(data.x.type(torch.long))
+        h_atm = self.embed_atm(data.x.type(torch.float))
         print(h_atm)
         print(h_atm.size())
         #h_bnd = self.embed_bnd(data.edge_attr)
